Creation methods were duplicated for backend. Get rid of them. Split tests into frontend and backend tests.
